Se agrega test basico

+++ src/main/java/com/supermercado/sistema_ventas/model/entities/Sucursal.java
@@ -20,6 +20,6 @@ public class Sucursal {
     private String rs;
     private String nomComercial;
     private String direccion;
-    private int RUC;
+    private Long RUC;
 
 }
